refactor(platform): fold the configurations form into its page
configurations-form.ts held a schema and one mapper used by a single page, so it
was a file boundary with nothing on the other side of it. Both now sit at the
bottom of index.tsx, and the configurationsForm wrapper object goes with it - that
grouping exists for cross-file callers, and toFormValues no longer has any.

telemetry-section.tsx needs ConfigurationsFormValues, which would make index and
telemetry-section import each other. It only ever used it as a type, so the import
is type-only and erases at compile time rather than forming a runtime cycle.

Both the form schema and UpdatePlatformConfigurationRequestBody restated
PlatformConfiguration's field names by hand, so a rename in shared would leave them
silently stale. They are picked from it now. The request body stays .partial(),
being a partial update; the form values do not, because the page renders a switch
per field and always holds a boolean.

Bumps shared, which nothing in this stack had done since main reached 0.159.0.
